pyproject.toml
[project]
name = "ai-marketing-agent"
version = "0.1.0"
description = "AI Marketing Agent for automated content creation and social media publishing"
requires-python = ">=3.12"
dependencies = [
    # Core
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",

    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "psycopg2-binary>=2.9",
    "alembic>=1.14",

    # Task Queue
    "celery[redis]>=5.4",
    "redis>=5.2",

    # AI / Content
    "anthropic>=0.52",
    "google-genai>=1.0",
    "jinja2>=3.1",
    "pillow>=11.0",

    # Platform APIs
    "httpx>=0.28",

    # Auth
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",

    # Notifications
    "slack-sdk>=3.33",

    # Utilities
    "python-multipart>=0.0.18",
    "boto3>=1.36",
    "tenacity>=9.0",
    "croniter>=5.0",
    "pytz>=2024.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.8",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
